+++ src/app/mod.rs
@@ -9,6 +9,8 @@ mod event_dispatcher;
 mod keybindings;
 pub mod model;
 mod snippets;
+mod spinner_ticking_port;
+mod status;
 pub(crate) mod storage;
 mod user_event;
 
@@ -19,5 +21,5 @@ pub(crate) use keybindings::KeybindingsConfig;
 pub(crate) use keybindings::SectionKeybindings;
 pub(crate) use keybindings::TisqKeyboundAction;
 pub(crate) use model::EditorId;
-pub(crate) use user_event::TisqEvent;
 pub(crate) use snippets::Snippet;
+pub(crate) use user_event::TisqEvent;