+++ src/app/user_event.rs
@@ -6,6 +6,7 @@ use super::{connection::DbResponse, EditorId};
 pub(crate) enum TisqEvent {
     TreeReloaded(SentTree),
     DbResponse(DbResponse),
+    SpinnerTick,
     EditorContentAdd(EditorId, String), // TODO: use attr instead of UserEvent
     EditorSnippetResolve {
         editor_id: EditorId,
@@ -19,6 +20,7 @@ pub(crate) enum TisqEvent {
 impl PartialEq for TisqEvent {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
+            (Self::SpinnerTick, Self::SpinnerTick) => true,
             (Self::TreeReloaded(_), Self::TreeReloaded(_)) => true,
             (Self::DbResponse(_), Self::DbResponse(_)) => true,
             (Self::EditorContentAdd(_, _), Self::EditorContentAdd(_, _)) => true,